EY-5142 forhåndsvarsel oms etteroppgjør (#7346)
* EY-5142 wip brevdata etteroppgjør

* EY-5142 wip brevdata etteroppgjør

* EY-5142 rette tester

* EY-5142 liten refaktorering

* EY-5142 fjerner referanser til gammelt brev

* EY-5142 la til lik dto som pensjonsbrev

* legge til gamle brevkoder

* hardkode rettsgebyr mot brev

* kun hente redigerbartinnhold hvis det skal være

* EY-5142 kun forhåndsvisning av brev

* EY-5142 la til forbehandling redigerbar

* EY-5142 endre rettsgebyr til int

* endre til kroner for rett formatering

+++ libs/etterlatte-beregning-model/src/main/kotlin/Etteroppgjoer.kt
@@ -24,7 +24,7 @@ data class BeregnetEtteroppgjoerResultatDto(
 data class EtteroppgjoerGrenseDto(
     val tilbakekreving: Double,
     val etterbetaling: Double,
-    val rettsgebyr: Double,
+    val rettsgebyr: Int,
     val rettsgebyrGyldigFra: LocalDate,
 )
